Switch from libcontainer users package to moby/sys/user
The libcontainer users package is deprecated in the main branch.
Podman's moving to vendor runc main now (lets us eliminate a
dependency that drags in 1mb of unnecessary code), so we need to
stop using the deprecated and removed runc package in favor of
the recommended replacement, github.com/moby/sys/user.

Signed-off-by: Matt Heon <mheon@redhat.com>